requirements.txt
aiosmtpd==1.4.6
annotated-types==0.7.0
anyio==4.9.0
atpublic==6.0.1
attrs==25.3.0
dnspython==2.7.0
fastapi==0.116.1
idna==3.10
motor==3.7.1
pydantic==2.11.7
pydantic_core==2.33.2
pymongo==4.13.2
python-dotenv==1.1.1
sniffio==1.3.1
starlette==0.47.2
typing-inspection==0.4.1
typing_extensions==4.14.1
email-validator==2.2.0

uvicorn[standard]==0.27.1


